Changed file format and moved file into compoenents
